+++ alacritty_terminal/src/vi_mode.rs
@@ -416,7 +416,7 @@ mod tests {
             padding_y: 0.0,
             dpr: 1.0,
         };
-        Term::new(&MockConfig::default(), &size, Mock)
+        Term::new(&MockConfig::default(), size, Mock)
     }
 
     #[test]